restructure authorize.ts; finish auth code flow
The authorization code flow now goes:
1. The client requests an authorization code from /_gridless/authorize
(essentially a session
code, as it's a token that lasts ten minutes) *with scopes*, client ID,
and `response_type=code`. The server returns this code in the JSON
body value of `code`.
2. The client sends the user to the browser, using /_gridless/authorize,
with the client ID, authorization code, and `grant_type=code`.
3. The user returns to the application, which should now request the
token from /_gridless/claimtoken, including its client ID and auth code.
The token will be in the JSON body value `token`.

Several security measures were added:
* The authorization code is verified, so that expirations and changes
are found and canceled.
* Before the token is given to the user, the token-in-progress value on
the user is cleared. The client can only get one token with this code.
* The redirect URI is validated against both the set redirect URIs
for the app and a hardcoded set of schemes and formats.
* The scopes are taken *from the token ONLY* when using the code flow.
That way the scopes are only set ONE TIME and cannot be changed to trick
the user into accepting a different set of scopes than the ones the app
actually claims.

Signed-off-by: Blake <bleonard252@rocketmail.com>